+++ CLIENT/src/MainSceneBuilderController.java
@@ -1,6 +1,9 @@
+import java.io.IOException;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
+import javafx.scene.control.Label;
 import javafx.scene.control.TextField;
+import client.MainTest;  // Assicurati che MainTest sia nel package corretto
 
 public class MainSceneBuilderController {
 
@@ -10,24 +13,35 @@ public class MainSceneBuilderController {
     @FXML
     private Button submitButton;
 
-    // Facoltativo, se vuoi lasciare MainTest nel controller
-    private MainTest mainTest;
+    @FXML
+    private Label feedbackLabel;
+
+    private MainTest mainTest;  // L'istanza di MainTest per la logica di connessione
 
     @FXML
-    public void onSubmitClicked() {
+    public void onSubmitClicked() throws ClassNotFoundException {
         String tableName = tableNameField.getText();
         System.out.println("Nome della tabella inserito: " + tableName);
 
-        // Qui puoi gestire la logica di invio della tabella tramite ClientApp
-        // Ad esempio, possiamo inviare il nome della tabella tramite una classe principale
+        if (tableName == null || tableName.trim().isEmpty()) {
+            feedbackLabel.setText("Nome della tabella non valido!");
+            return;
+        }
 
-        // Se vuoi usare mainTest dal controller (ma se non lo vuoi puoi farlo direttamente in ClientApp)
         if (mainTest != null) {
-            mainTest.sendTableName(tableName); // invio al server tramite MainTest
+            try {
+                mainTest.sendTableName(tableName);
+                feedbackLabel.setText("Tabella inviata correttamente!");
+            } catch (IOException e) {
+                feedbackLabel.setText("Errore nella connessione al server!");
+                e.printStackTrace();
+            }
+        } else {
+            feedbackLabel.setText("Errore: MainTest non inizializzato!");
         }
     }
 
-    // Opzionale, se vuoi avere un metodo di iniezione per MainTest
+    // Metodo di iniezione per MainTest
     public void setMainTest(MainTest mainTest) {
         this.mainTest = mainTest;
     }
